+++ pypsa/contingency.py
@@ -37,7 +37,7 @@ import pandas as pd
 
 import collections
 
-from .pf import calculate_PTDF
+from .pf import calculate_PTDF, _as_snapshots
 
 from .opt import l_constraint
 
@@ -90,14 +90,11 @@ def network_lpf_contingency(network, snapshots=None, branch_outages=None):
     ----------
     snapshots : list-like|single snapshot
         A subset or an elements of network.snapshots on which to run
-        the power flow, defaults to [now]
+        the power flow, defaults to network.snapshots
         NB: currently this only works for a single snapshot
     branch_outages : list-like
         A list of passive branches which are to be tested for outages.
         If None, it's take as all network.passive_branches_i()
-    now : object
-        Deprecated: A member of network.snapshots on which to run the
-        power flow, defaults to network.now
 
     Returns
     -------
@@ -109,8 +106,9 @@ def network_lpf_contingency(network, snapshots=None, branch_outages=None):
     from .components import passive_branch_components
 
     if snapshots is None:
-        snapshot = network.now
-    elif isinstance(snapshots, collections.Iterable):
+        snapshots = network.snapshots
+
+    if isinstance(snapshots, collections.Iterable):
         logger.warning("Apologies LPF contingency, this only works for single snapshots at the moment, taking the first snapshot.")
         snapshot = snapshots[0]
     else:
@@ -167,7 +165,7 @@ def network_sclopf(network,snapshots=None,branch_outages=None,solver_name="glpk"
     Parameters
     ----------
     snapshots : list or index slice
-        A list of snapshots to optimise, must be a subset of network.snapshots, defaults to network.now
+        A list of snapshots to optimise, must be a subset of network.snapshots, defaults to network.snapshots
     branch_outages : list-like
         A list of passive branches which are to be tested for outages.
         If None, it's take as all network.passive_branches_i()
@@ -193,8 +191,7 @@ def network_sclopf(network,snapshots=None,branch_outages=None,solver_name="glpk"
     if not skip_pre:
         network.determine_network_topology()
 
-    if snapshots is None:
-        snapshots = [network.now]
+    snapshots = _as_snapshots(network, snapshots)
 
     passive_branches = network.passive_branches()