+++ pypsa/pf.py
@@ -46,7 +46,7 @@ from .descriptors import get_switchable_as_dense, allocate_series_dataframes
 
 def _as_snapshots(network, snapshots):
     if snapshots is None:
-        snapshots = [network.now]
+        snapshots = network.snapshots
     if (isinstance(snapshots, six.string_types) or
         not isinstance(snapshots, (collections.Sequence, pd.Index))):
         return pd.Index([snapshots])
@@ -116,7 +116,7 @@ def network_pf(network, snapshots=None, skip_pre=False, x_tol=1e-6, use_seed=Fal
     ----------
     snapshots : list-like|single snapshot
         A subset or an elements of network.snapshots on which to run
-        the power flow, defaults to [now]
+        the power flow, defaults to network.snapshots
     skip_pre: bool, default False
         Skip the preliminary steps of computing topology, calculating dependent values and finding bus controls.
     x_tol: float
@@ -171,7 +171,7 @@ def sub_network_pf(sub_network, snapshots=None, skip_pre=False, x_tol=1e-6, use_
     ----------
     snapshots : list-like|single snapshot
         A subset or an elements of network.snapshots on which to run
-        the power flow, defaults to [now]
+        the power flow, defaults to network.snapshots
     skip_pre: bool, default False
         Skip the preliminary steps of computing topology, calculating dependent values and finding bus controls.
     x_tol: float
@@ -366,7 +366,7 @@ def network_lpf(network, snapshots=None, skip_pre=False):
     ----------
     snapshots : list-like|single snapshot
         A subset or an elements of network.snapshots on which to run
-        the power flow, defaults to [now]
+        the power flow, defaults to network.snapshots
     skip_pre: bool, default False
         Skip the preliminary steps of computing topology, calculating
         dependent values and finding bus controls.
@@ -875,7 +875,7 @@ def sub_network_lpf(sub_network, snapshots=None, skip_pre=False):
     ----------
     snapshots : list-like|single snapshot
         A subset or an elements of network.snapshots on which to run
-        the power flow, defaults to [now]
+        the power flow, defaults to network.snapshots
     skip_pre: bool, default False
         Skip the preliminary steps of computing topology, calculating
         dependent values and finding bus controls.