+++ test/test_lpf_against_pypower.py
@@ -55,7 +55,7 @@ def test_pypower_case():
 
     #compare generator dispatch
 
-    p_pypsa = network.generators_t.p.loc[network.now].values
+    p_pypsa = network.generators_t.p.loc["now"].values
     p_pypower = results_df['gen']["p"].values
 
     np.testing.assert_array_almost_equal(p_pypsa,p_pypower)
@@ -66,6 +66,6 @@ def test_pypower_case():
         pnl = getattr(network,item + "_t")
 
         for si in ["p0","p1"]:
-            si_pypsa = getattr(pnl,si).loc[network.now].values
+            si_pypsa = getattr(pnl,si).loc["now"].values
             si_pypower = results_df['branch'][si][df.original_index].values
             np.testing.assert_array_almost_equal(si_pypsa,si_pypower)